requirements.txt
openai==0.28
requests==2.31.0
matplotlib==3.8.4
psycopg2-binary==2.9.9